+++ src/plafosim/simulator.py
@@ -193,7 +193,7 @@ class Simulator:
                     if self._change_lane(vehicle.vid, target_lane):
                         # log lane change
                         with open(self._result_base_filename + '_vehicle_changes.csv', 'a') as f:
-                            f.write("%d,%d,%d,%d,%d,%d,%s\n" % (self.step, vehicle.vid, vehicle.position, source_lane, target_lane, vehicle.speed, "speedGain"))
+                            f.write("%d,%d,%d,%d,%d,%f,%s\n" % (self.step, vehicle.vid, vehicle.position, source_lane, target_lane, vehicle.speed, "speedGain"))
             else:
                 if vehicle.lane > 0:
                     source_lane = vehicle.lane
@@ -201,7 +201,7 @@ class Simulator:
                     if self._change_lane(vehicle.vid, target_lane):
                         # log lane change
                         with open(self._result_base_filename + '_vehicle_changes.csv', 'a') as f:
-                            f.write("%d,%d,%d,%d,%d,%d,%s\n" % (self.step, vehicle.vid, vehicle.position, source_lane, target_lane, vehicle.speed, "keepRight"))
+                            f.write("%d,%d,%d,%d,%d,%f,%s\n" % (self.step, vehicle.vid, vehicle.position, source_lane, target_lane, vehicle.speed, "keepRight"))
 
     def adjust_speeds(self):
         """Do speed adjustments for all vehicles"""
@@ -296,14 +296,14 @@ class Simulator:
             penetration_rate: float,
             depart_interval: int,
             arrival_interval: int,
-            max_speed: int,
+            max_speed: float,
             random_depart_position: bool,
             random_depart_lane: bool,
-            desired_speed: int,
+            desired_speed: float,
             random_desired_speed: bool,
             speed_variation: float,
-            min_desired_speed: int,
-            max_desired_speed: int,
+            min_desired_speed: float,
+            max_desired_speed: float,
             random_depart_speed: bool,
             depart_desired: bool,
             depart_method: str,
